requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
rembg==2.0.67
onnxruntime==1.24.2
pillow==11.3.0
python-multipart==0.0.20
